requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.9.0
pydantic-settings>=2.6.0
groq>=0.13.0
pillow>=10.4.0
python-dotenv>=1.0.0
httpx>=0.27.0
opencv-python-headless>=4.10.0.84,<4.12.0
numpy>=1.21.6,<1.25.0
setuptools
wheel